doc/src/requirements.txt
requests
Werkzeug==0.16.1
flask-lambda-python36==0.1.0
flask-restplus==0.13.0
pyyaml>=5.3.1